Switch server/client
+ Reference .proto folder
