+++ Kernel/Connection/Protos/greet.proto
@@ -1,21 +1,21 @@
 syntax = "proto3";
 
-option csharp_namespace = "Distributor";
+option csharp_namespace = "Responder";
 
 package greet;
 
 // The greeting service definition.
 service Greeter {
-	// Sends a greeting
-	rpc SayHello (HelloRequest) returns (HelloReply);
+  // Sends a greeting
+  rpc SayHello (HelloRequest) returns (HelloReply);
 }
 
 // The request message containing the user's name.
 message HelloRequest {
-	string name = 1;
+  string name = 1;
 }
 
 // The response message containing the greetings.
 message HelloReply {
-	string message = 1;
+  string message = 1;
 }
