+++ Kernel/Kernel.sln
@@ -5,9 +5,11 @@ VisualStudioVersion = 16.0.30523.141
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TaskManager", "TaskManager\TaskManager.csproj", "{9D57CA06-A033-4F14-AADD-2BA72415D7D2}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Distributor", "Distributor\Distributor.csproj", "{E69E0EB1-38A0-4CC2-9424-581E8AAE0896}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Responder", "Responder\Responder.csproj", "{784F8AD7-12D1-4408-B529-E015920A6337}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Responder", "Responder\Responder.csproj", "{50BDA7A1-6F28-414F-B4FE-E3C58E515D0E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Distributor", "Distributor\Distributor.csproj", "{2D643C14-A60F-4992-BC3B-AA7908720337}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Connection", "Connection\Connection.csproj", "{48552BF6-3EA1-408C-9DEA-3E38F9BA9FF5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -19,14 +21,18 @@ Global
 		{9D57CA06-A033-4F14-AADD-2BA72415D7D2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9D57CA06-A033-4F14-AADD-2BA72415D7D2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9D57CA06-A033-4F14-AADD-2BA72415D7D2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E69E0EB1-38A0-4CC2-9424-581E8AAE0896}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E69E0EB1-38A0-4CC2-9424-581E8AAE0896}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E69E0EB1-38A0-4CC2-9424-581E8AAE0896}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E69E0EB1-38A0-4CC2-9424-581E8AAE0896}.Release|Any CPU.Build.0 = Release|Any CPU
-		{50BDA7A1-6F28-414F-B4FE-E3C58E515D0E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{50BDA7A1-6F28-414F-B4FE-E3C58E515D0E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{50BDA7A1-6F28-414F-B4FE-E3C58E515D0E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{50BDA7A1-6F28-414F-B4FE-E3C58E515D0E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{784F8AD7-12D1-4408-B529-E015920A6337}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{784F8AD7-12D1-4408-B529-E015920A6337}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{784F8AD7-12D1-4408-B529-E015920A6337}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{784F8AD7-12D1-4408-B529-E015920A6337}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2D643C14-A60F-4992-BC3B-AA7908720337}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2D643C14-A60F-4992-BC3B-AA7908720337}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2D643C14-A60F-4992-BC3B-AA7908720337}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2D643C14-A60F-4992-BC3B-AA7908720337}.Release|Any CPU.Build.0 = Release|Any CPU
+		{48552BF6-3EA1-408C-9DEA-3E38F9BA9FF5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{48552BF6-3EA1-408C-9DEA-3E38F9BA9FF5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{48552BF6-3EA1-408C-9DEA-3E38F9BA9FF5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{48552BF6-3EA1-408C-9DEA-3E38F9BA9FF5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
