+++ Kernel/Responder/Properties/launchSettings.json
@@ -1,6 +1,6 @@
 ﻿{
   "profiles": {
-    "Distributor": {
+    "Responder": {
       "commandName": "Project",
       "launchBrowser": false,
       "applicationUrl": "https://localhost:5001",
