+++ Kernel/Responder/Startup.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-namespace Distributor
+namespace Responder
 {
     public class Startup
     {
