+++ Kernel/TaskManager/Core/AdjacencyMatrix.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace TaskManager
+{
+    class AdjacencyMatrix
+    {
+    }
+}
