+++ Kernel/TaskManager/Core/Route.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace ODC.TaskManager
+{
+    class Route
+    {
+    }
+}
